removed page1&2, refactored controller for eq and home
